+++ spec/server/process_spec.rb
@@ -204,6 +204,8 @@ describe "Sensu::Server::Process" do
   end
 
   it "can process results with flap detection" do
+    FileUtils.rm_rf("/tmp/sensu_event")
+    expect(File.exists?("/tmp/sensu_event")).to eq(false)
     async_wrapper do
       @server.setup_redis do
         redis.flushdb do
@@ -228,11 +230,16 @@ describe "Sensu::Server::Process" do
                     result[:check][:low_flap_threshold] = 5
                     result[:check][:high_flap_threshold] = 20
                     result[:check][:status] = 0
+                    result[:check][:handler] = "file"
                     @server.process_check_result(result)
                   end
                   timer(1) do
                     redis.hexists("events:i-424242", "test") do |exists|
                       expect(exists).to be(false)
+                      expect(File.exists?("/tmp/sensu_event")).to eq(true)
+                      event_file_json = IO.read("/tmp/sensu_event")
+                      event = Sensu::JSON.load(event_file_json)
+                      expect(event[:action]).to eq("resolve")
                       async_done
                     end
                   end
@@ -277,10 +284,10 @@ describe "Sensu::Server::Process" do
   end
 
   it "can expire silenced entries on event resolution" do
+    FileUtils.rm_rf("/tmp/sensu_event")
+    expect(File.exists?("/tmp/sensu_event")).to eq(false)
     async_wrapper do
       @server.setup_redis do
-        FileUtils.rm_rf("/tmp/sensu_event")
-        expect(File.exists?("/tmp/sensu_event")).to eq(false)
         redis.flushdb do
           redis.set("client:i-424242", Sensu::JSON.dump(client_template)) do
             silenced_info = {
@@ -699,7 +706,7 @@ describe "Sensu::Server::Process" do
       check = check_template
       check[:cron] = "* * * * *"
       @server.schedule_checks([check])
-      expect(@server.instance_variable_get(:@timers)[:leader].size).to eq(1)
+      expect(@server.instance_variable_get(:@timers)[:tasks][:check_request_publisher].size).to eq(1)
       async_done
     end
   end
@@ -857,15 +864,16 @@ describe "Sensu::Server::Process" do
     end
   end
 
-  it "can be the leader and resign" do
+  it "can be responsible for all tasks and relinquish them" do
     async_wrapper do
       @server.setup_connections do
         redis.flushdb do
-          @server.request_leader_election
-          timer(1) do
-            expect(@server.is_leader).to be(true)
-            @server.resign_as_leader
-            expect(@server.is_leader).to be(false)
+          expect(@server.tasks).to be_empty
+          @server.setup_task_elections(0)
+          timer(2) do
+            expect(@server.tasks).to eq(["check_request_publisher", "client_monitor", "check_result_monitor"])
+            @server.relinquish_tasks
+            expect(@server.tasks).to be_empty
             async_done
           end
         end
@@ -873,21 +881,22 @@ describe "Sensu::Server::Process" do
     end
   end
 
-  it "can be the only leader" do
+  it "can be responsible for one or more tasks" do
     async_wrapper do
-      server1 = @server.clone
-      server2 = @server.clone
+      server1 = Sensu::Server::Process.new(options)
+      server2 = Sensu::Server::Process.new(options)
       server1.setup_connections do
         server2.setup_connections do
           redis.flushdb do
-            lock_timestamp = (Time.now.to_f * 1000).to_i - 60000
-            redis.set("lock:leader", lock_timestamp) do
-              server1.setup_leader_monitor
-              server2.setup_leader_monitor
-              timer(3) do
-                expect([server1.is_leader, server2.is_leader].uniq.size).to eq(2)
-                async_done
-              end
+            expect(server1.tasks).to be_empty
+            expect(server2.tasks).to be_empty
+            server1.setup_task_elections(1)
+            server2.setup_task_elections(1)
+            timer(2) do
+              expect(server1.tasks).not_to be_empty
+              expect(server2.tasks).not_to be_empty
+              expect(server1.tasks - server2.tasks).to eq(server1.tasks)
+              async_done
             end
           end
         end
@@ -906,6 +915,9 @@ describe "Sensu::Server::Process" do
                 expect(server_json).to be_kind_of(String)
                 server = Sensu::JSON.load(server_json)
                 expect(server[:id]).to eq(@server.server_id)
+                expect(server[:tasks]).to be_kind_of(Array)
+                expect(server[:sensu][:version]).to eq(Sensu::VERSION)
+                expect(server[:sensu][:settings][:hexdigest]).to be_kind_of(String)
                 expect(server[:timestamp]).to be_within(5).of(Time.now.to_i)
                 async_done
               end